Revert changes to SNP naming (#28)
As part of the changes to combining probes at the same locus, I switched from using the 'snp name' column to identify probes to 'illumina id' with the idea that multiple probes could have the same snp name. This ended up not working anyway, so I had to make snp names unique. It turns out that in production data the illumina id column is not always present, so data processing is failing in production.

This PR reverts to the old behaviour where probes are identified by snp name

Ideally we'd write some new tests to reflect the problem encountered in production to prevent this from happening again in future, but I have neither the time nor the skills to figure that out right now. This should unblock Summit data processing for now at least